+++ internal/collect/request.go
@@ -4,6 +4,7 @@ import (
 	"crypto/md5"
 	"encoding/hex"
 	"errors"
+	"github.com/xjh-creator/reptile/internal/collector"
 	"regexp"
 	"sync"
 	"time"
@@ -26,6 +27,7 @@ type Task struct {
 	VisitedLock sync.Mutex
 	Rule        RuleTree
 	Fetcher     Fetcher
+	Store       collector.Store
 }
 
 
@@ -42,6 +44,7 @@ type Request struct {
 	Depth    int64
 	Priority int64
 	RuleName string
+	TmpData  *Temp
 }
 
 // ParseResult 爬取后获取的数据
@@ -82,6 +85,7 @@ func (c *Context) ParseJSReg(name string, reg string) ParseResult {
 				RuleName: name,
 			})
 	}
+
 	return result
 }
 
@@ -96,5 +100,22 @@ func (c *Context) OutputJS(reg string) ParseResult {
 	result := ParseResult{
 		Items: []interface{}{c.Req.Url},
 	}
+
 	return result
+}
+
+func (c *Context) GetRule(ruleName string) *Rule {
+
+	return c.Req.Task.Rule.Trunk[ruleName]
+}
+
+func (c *Context) Output(data interface{}) *collector.OutputData {
+	res := &collector.OutputData{}
+	res.Data = make(map[string]interface{})
+	res.Data["Rule"] = c.Req.RuleName
+	res.Data["Data"] = data
+	res.Data["Url"] = c.Req.Url
+	res.Data["Time"] = time.Now().Format("2006-01-02 15:04:05")
+
+	return res
 }