+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## feat:reptile struct of data
+1. 爬取结构化的数据
+
 ##  feat:add parse js
 1. 动态规则引擎
 2. 实现动态规则的引擎，可 Javascript 虚拟机，因为使用 JS 操作网页有天然的优势